feat, bugfix!: Fixed Result type, implemented JSON conversion helpers
The Result type did not have a way to have a variadic for to be able to return from functions which specified `Result[Tally]` with a `Result[string]` for a Result constructed with `Err(":3")`. Accordingly, there are now two internal values for a Result type, with the error value always being a string to suade the compiler.

Additionally, there are now two functions new to the `Tally` type. You can (attempt to) construct (deserialize) a `Tally` from JSON using the `TallyFromJson` function, and you can serialize it back into a JSON string. This is however seperate from the `String` conversion function specified in #1.

Completes #9